pyproject.toml
[project]
name = "cv-mcp"
version = "0.1.0"
description = "MCP server exposing Francesco Foresta's CV as a queryable knowledge source"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Francesco Foresta", email = "fr.foresta@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "fastapi",
    "uvicorn",
    "pydantic",
    "anthropic",
    "openai>=2.35.0",
    "google-genai>=1.75.0",
]

[project.scripts]
cv-mcp-server = "cv_mcp.server:main"

[dependency-groups]
dev = [
    "pytest",
    "httpx",
]

[build-system]
requires = ["uv_build>=0.11.9,<0.12.0"]
build-backend = "uv_build"
